+++ ambient_download/src/query.rs
@@ -1,4 +1,5 @@
 /// The commands issued from the CLI
+#[allow(clippy::module_name_repetitions)]
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
 pub enum QueryType {
     /// Get the device information.
@@ -6,6 +7,12 @@ pub enum QueryType {
 
     /// Get the weather information.
     GetWeather,
+
+    /// Output the local time zone information
+    GetTimezone,
+
+    /// Create a new, empty config file
+    NewConfig,
 
     /// Print the help message.
     #[default]
@@ -18,6 +25,8 @@ impl QueryType {
         match cli_args.subcommand_name() {
             Some("device") => Self::GetDeviceInfo,
             Some("weather") => Self::GetWeather,
+            Some("timezone") => Self::GetTimezone,
+            Some("newconfig") => Self::NewConfig,
             _ => Self::Help,
         }
     }