+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+### Feat
+
+- Blank config file generation
+
+## [0.5.0] - 2024-05-18
+
 ### Chore
 
 - Delete ambient_download.toml
@@ -15,6 +21,11 @@ All notable changes to this project will be documented in this file.
 ### Feat
 
 - Create a blank config file
+- Time zone info in main app
+
+### Fix
+
+- Tz_detect create new blank config if not found
 
 ### Refactor